Feat: Implemented BaseCalculator with a BigDecimal type. Implemented an expression parser, the logarithm parser and the trigonometry parser. Implemented rules for BaseCalculator. Fixed: displaying the value in ValueTranslator.

+++ app/src/main/java/com/cobaltumapps/simplecalculator/v15/calculator/components/display/DisplayExpressionController.kt
@@ -1,7 +1,7 @@
 package com.cobaltumapps.simplecalculator.v15.calculator.components.display
 
-import com.cobaltumapps.simplecalculator.v15.calculator.components.display.interfaces.DisplayExpressionControl
 import com.cobaltumapps.simplecalculator.data.calculator.enums.AngleMode
+import com.cobaltumapps.simplecalculator.v15.calculator.components.display.interfaces.DisplayExpressionControl
 import com.cobaltumapps.simplecalculator.v15.calculator.services.memory.ValueTranslator
 
 open class DisplayExpressionController(
@@ -10,22 +10,18 @@ open class DisplayExpressionController(
     private var displayLogger = DisplayExpressionLogger()
     private val displayValueTranslator = ValueTranslator()
 
-    // Logic
     private var canAnimateResult = false
 
-    // Установка значения типа угла
     override fun setAngleField(angleMode: AngleMode) {
         displayComponent.setAngleField(angleMode)
         displayLogger.setAngleField(angleMode)
     }
 
-    // Установка значения в поле памяти
     override fun setMemoryField(memoryValue: Number) {
         displayComponent.setMemoryField(memoryValue)
         displayLogger.setMemoryField(memoryValue)
     }
 
-    // Установка значения в поле выражения
     override fun setExpressionField(newExpression: String) {
         displayComponent.setExpressionField(newExpression)
         if (canAnimateResult) {
@@ -35,11 +31,8 @@ open class DisplayExpressionController(
         displayLogger.setExpressionField(newExpression)
     }
 
-    // Установка значение в поле результата
     override fun setResultField(newResult: String) {
-        displayComponent.setResultField(
-            displayValueTranslator.translate(newResult.toDouble())
-        )
+        displayComponent.setResultField(displayValueTranslator.translate(newResult))
 
         if (!canAnimateResult) {
             canAnimateResult = true
@@ -48,13 +41,11 @@ open class DisplayExpressionController(
             displayLogger.setResultField(newResult)
     }
 
-    // Очистка выражения
     override fun clearExpressionField() {
         displayComponent.setExpressionField("")
         displayLogger.clearExpressionField()
     }
 
-    // Очистка результата
     override fun clearResultField() {
         displayComponent.displayAnimator.playHiddenResultAnim {
             displayComponent.setResultField("")
